+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 * Respect `nil` value when provided, by skipping model attribute value.
+* Add support for Rails 6.1, drop support for Rails < 5.2.
 * Move CI to GitHub Actions.
 
 ## 0.7.0